Add cast to right type for printf() field precision specifier
Before, I was getting a warning:

sysfs-poll.c:34:12: warning: field precision specifier '.*' expects argument of type 'int', but argument 2 has type 'ssize_t' {aka 'long int'} [-Wformat=]

This typecast should be safe from overflow because the value of
read_bytes is constrained by the size of readbuf[], and it's previously
checked not to be negative.

+++ sysfs-poll.c
@@ -31,7 +31,7 @@ int main(int argc, char * argv[])
 		perror(argv[1]);
 		exit(3);
 	}
-	printf("%.*s", read_bytes, readbuf);
+	printf("%.*s", (int)read_bytes, readbuf);
 
 	pfd.fd = attr_fd;
 	pfd.events = POLLERR | POLLPRI;
@@ -46,7 +46,7 @@ int main(int argc, char * argv[])
 				perror(argv[1]);
 				exit(4);
 			}
-			printf("%.*s", read_bytes, readbuf);
+			printf("%.*s", (int)read_bytes, readbuf);
 		}
 	}
 	return 0;
